# backend/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Referenda",
        "description": "Spanish Pleno del Congreso de los Diputados unofficial API\n\nToken needed for certain operations.\n\nPlease ask for it at webmaster@referenda.es\n"
    },
    "host": "localhost:3443",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/laws": {
            "get": {
                "description": "Returns all laws being debated or debated in Congress",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "results",
                        "in": "query",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all laws",
                        "schema": {
                            "title": "Law",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Law"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new law being debated in Congress",
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "law",
                        "in": "body",
                        "required": true,
                        "description": "The law JSON to be created",
                        "schema": {
                            "$ref": "#/definitions/LawExtended"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Create a new law"
                    }
                }
            }
        },
        "/laws/{slug}": {
            "get": {
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the law requested",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the law",
                        "schema": {
                            "$ref": "#/definitions/LawWithComments"
                        }
                    }
                }
            },
            "put": {
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "The access token needed for updating a law",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the law to be modified",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sends the updated law",
                        "schema": {
                            "$ref": "#/definitions/Law"
                        }
                    }
                }
            },
            "delete": {
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "The access token needed for deleting a law",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the law to be deleted",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "law successfuly deleted"
                    }
                }
            }
        },
        "/laws/{slug}/votes": {
            "post": {
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "The access token needed for voting a law",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the law to be voted",
                        "required": true
                    },
                    {
                        "name": "vote",
                        "in": "body",
                        "description": "vote to be added to the law",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Vote"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "added vote to the law",
                        "schema": {
                            "$ref": "#/definitions/LawExtended"
                        }
                    }
                }
            }
        },
        "/laws/{slug}/comments": {
            "get": {
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the requested law",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "comments posted by the community for the requested law",
                        "schema": {
                            "title": "comments",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Comment"
                            }
                        }
                    }
                }
            },
            "post": {
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "The access token needed for posting a comment",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the law to be commented",
                        "required": true
                    },
                    {
                        "name": "comment",
                        "in": "body",
                        "description": "comment to be added to the law",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Comment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "added comment to the low"
                    }
                }
            },
            "delete": {
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "The access token needed for deleting law comments",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the law having the comments to be deleted",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted all comments from the low"
                    }
                }
            }
        },
...
        "/laws/{slug}/comments/{commentId}/votes": {
            "post": {
                "parameters": [
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "The access token needed for voting a comment",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "type": "string",
                        "description": "slug of the law with the comment to be voted",
                        "required": true
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "type": "string",
                        "description": "commentId of the comment to be updated",
                        "required": true
                    },
                    {
                        "name": "vote",
                        "in": "body",
                        "description": "vote to be added to the comment",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Vote"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "added vote to comment of the low"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Law": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "law_type": {
                    "description": "type of the law",
                    "type": "string"
                },
                "institution": {
                    "description": "institution of party behind the law",
                    "type": "string"
                },
                "tier": {
                    "description": "importance of the law. 1 to 3 levels",
                    "type": "string"
                },
                "headline": {
                    "description": "short headline of the law",
                    "type": "string"
                },
                "slug": {
                    "description": "very short unique headline of the law",
                    "type": "string"
                },
                "short_description": {
                    "description": "brief description of the law",
                    "type": "string"
                },
                "long_description": {
                    "description": "long description of the law",
                    "type": "string"
                },
                "link": {
                    "description": "official link in the Congress online publications",
                    "type": "string"
                },
                "pub_date": {
                    "description": "date the law is publicated in referenda.es",
                    "type": "string"
                },
                "vote_start": {
                    "description": "date the law can be votedin referenda.es",
                    "type": "string"
                },
                "vote_end": {
                    "description": "date the law can no longer be voted in referenda.es",
                    "type": "string"
                },
                "positive": {
                    "description": "positive votes received by the law in referenda.es",
                    "type": "integer",
                    "format": "int32"
                },
                "negative": {
                    "description": "negative votes received by the law in referenda.es",
                    "type": "integer",
                    "format": "int32"
                },
                "abstention": {
                    "description": "abstention votes received by the law in referenda.es",
                    "type": "integer",
                    "format": "int32"
                },
                "official_positive": {
                    "description": "positive votes received by the law in the Congress",
                    "type": "integer",
                    "format": "int32"
                },
                "official_negative": {
                    "description": "negative votes received by the law in the Congress",
                    "type": "integer",
                    "format": "int32"
                },
                "official_abstention": {
                    "description": "abstention votes received by the law in the Congress",
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "Vote": {
            "type": "object",
            "properties": {
                "vote": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
    }
}